# docs/swagger.yaml
definitions:
...
  model.RedEnvelopeType:
    enum:
    - fixed
    - random
    type: string
    x-enum-varnames:
    - RedEnvelopeTypeFixed
    - RedEnvelopeTypeRandom
...
  redenvelope.ClaimRequest:
    properties:
      id:
        example: "0"
        type: string
    required:
    - id
    type: object
  redenvelope.CreateRequest:
    properties:
      cover_upload_id:
        example: "0"
        type: string
      greeting:
        maxLength: 100
        type: string
      heterotypic_upload_id:
        example: "0"
        type: string
      pay_key:
        maxLength: 10
        type: string
      total_amount:
        type: number
      total_count:
        minimum: 1
        type: integer
      type:
        allOf:
        - $ref: '#/definitions/model.RedEnvelopeType'
        enum:
        - fixed
        - random
    required:
    - pay_key
    - total_amount
    - total_count
    - type
    type: object
  redenvelope.ListRequest:
    properties:
      page:
        minimum: 1
        type: integer
      page_size:
        maximum: 100
        minimum: 1
        type: integer
      type:
        enum:
        - sent
        - received
        type: string
    required:
    - page
    - page_size
    type: object
...
  util.ResponseAny:
    properties:
      data: {}
      error_msg:
        example: ""
        type: string
    type: object
info:
  contact: {}
  title: LINUX DO Credit
  version: 1.0.0
paths:
...
  /api/v1/redenvelope/{id}:
    get:
      parameters:
      - description: 红包ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ResponseAny'
      tags:
      - redenvelope
  /api/v1/redenvelope/claim:
    post:
      consumes:
      - application/json
      parameters:
      - description: 领取红包请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/redenvelope.ClaimRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ResponseAny'
      tags:
      - redenvelope
...
  /api/v1/redenvelope/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建红包请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/redenvelope.CreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ResponseAny'
      tags:
      - redenvelope
  /api/v1/redenvelope/list:
    post:
      consumes:
      - application/json
      parameters:
      - description: 列表请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/redenvelope.ListRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ResponseAny'
      tags:
      - redenvelope
...
swagger: "2.0"